feat: components 분리
