+++ src/basic/components/index.ts
@@ -1,0 +1,5 @@
+export { Header } from './Header';
+export { ProductCard } from './ProductCard';
+export { ProductList } from './ProductList';
+export { Cart } from './Cart';
+export { NotificationList } from './NotificationList';
